+++ Server/accounts/urls.py
@@ -1,9 +1,8 @@
 from django.urls import path
-from .views import HelloWorldView, GoogleLogin, github_callback, GithubLogin, KakaoLogin, kakao_callback, \
+from .views import GoogleLogin, github_callback, GithubLogin, KakaoLogin, kakao_callback, \
     google_callback
 
 urlpatterns = [
-    path('hello/', HelloWorldView.as_view(), name='hello_world'),
     path('dj-rest-auth/google/', google_callback.as_view(), name="google_login"),
     path('dj-rest-auth/github/', github_callback.as_view(), name="github_callback"),
     path('dj-rest-auth/kakao/', kakao_callback.as_view(), name="kakao_callback"),
