+++ Server/profiles/serializers.py
@@ -14,18 +14,3 @@ class GroupCreationSerializer(serializers.ModelSerializer):
     class Meta:
         model = Group
         fields = '__all__'
-
-class IsOwnerOrReadOnly(permissions.BasePermission):
-    def has_object_permission(self, request, view, obj):
-        if request.user.is_authenticated:
-            if request.user.is_staff:
-                return True
-            elif request.method in permissions.SAFE_METHODS:
-                return True
-            elif hasattr(obj, 'user_pk'):
-                return obj.user_pk == request.user.id
-            elif obj.__class__ == get_user_model():
-                return obj.user_pk == request.user.id
-            return False
-        else:
-            return False
